add a separate php file for functions, add docstrings, hasn't replaced nested if yet
